wproc: Make print_input() a bit more clever
Since we want to make sure things get written, read and parsed properly
when travelling from worker to master we need to print out the vars
one by one, and once for each request. With this patch, we get just
that and can rest assured that communication works exactly as it's
supposed to.

Signed-off-by: Andreas Ericsson <ae@op5.se>