lib/worker: Avoid trying to run jobs when we can't parse command
This is two-fold; We need to be able to parse the worker-command (ie,
the key/value vector we're sent from the master) as well as the
command line (ie, the part we're going to ship off to execvp()).

If either fails, we used to crash out, but with this patch we'll keep
on chugging along. In order to provide proper error messages to our
master, we introduce the "job_error()" function, which responds to the
master with a replica of the request and an error message in case
anything goes wrong.

Signed-off-by: Andreas Ericsson <ae@op5.se>