lib/worker: Make the child process list double-ended
We'll want to know how long we can poll our children before we run
into a timeout, so we need to keep a tail of the list up to date as
well. Since we'll be running commands that in 99.99999% of all cases
have the exact same timeout, handling the child list will be very
cheap, as only "add to end", "peek at first" and "remove any one" all
have an asymptotic complexity of O(1).

Signed-off-by: Andreas Ericsson <ae@op5.se>