lib/iocache: Make iocache_read() read to the right location
Previously we used to have iocache_read() read to the location
pointed to by the ioc_offset variable, which meant we would overwrite
previously read but not used data in case all read data wasn't used
in one go. Since we don't have any callers that use the iocache()
layer in that manner, we didn't trigger the bug, but it needs fixing
even so.

Signed-off-by: Andreas Ericsson <ae@op5.se>