lib/iocache: Properly discard already checked data in iocache_use_delim()
Previously we used to run memchr() with iocache_available() - delim_len
as marker for how much memory to scan, but since we're advancing the
pointer each iteration we need to shrink the scanned area as well.

We never triggered this bug, since our messages are always nul-delimited
and the top-most loop condition bailed us out before we could hit it,
but it needs fixing even so.

Signed-off-by: Andreas Ericsson <ae@op5.se>