lib/worker: Don't modify the I/O cache directly
Since we're using a delimiter of nul bytes and always keep track of
the length of the buffer, we don't have to poke around in it manually.
The manual poking could also lead to us accessing beyond the buffer,
since buflen is already updated but we used that and added ioc_ret to
it to find the end of the buffer. In case we read more than half the
I/O cache in one go, that would put us beyond the end of the buffer.

Signed-off-by: Andreas Ericsson <ae@op5.se>