base/workers: Add wproc_notify()
Notifications are one of the major irks when it comes to keeping
latency down, since they're currently run asynchronously, so they
block all checks from happening while they're executing. In small
installations, that might not be a large problem, but for larger
ones it can be quite a hassle, as potentially many contacts or very
slow-running scripts that integrate Nagios with other pieces of
software can cause major latency increases.

Signed-off-by: Andreas Ericsson <ae@op5.se>