lib/iocache: Also update bookkeeping variables when moving data
We used to just move the data and not update the bookkeeping
variables, which lead us to the very deadlock the iocache_move_data()
function was invented to prevent, namely ioc_buflen being equal to
ioc_bufsize and us issuing zero size reads.

This patch fixes it so the function once again does what it's
supposed to do.

Signed-off-by: Andreas Ericsson <ae@op5.se>