base/workers: Truly destroy the job in destroy_job()
Previously we didn't even mark the job-slot as being unused. Since
our pretty uncaring code that assigned job-slots was so sloppy it
didn't care about that, the only bad thing that happened was that
we leaked some memory. This patch fixes the leak. The sloppy code is
allowed to linger a while longer.

Signed-off-by: Andreas Ericsson <ae@op5.se>