base/workers: Take more care not to overwrite old jobs when starting new
Previously we used to ignore the fact that a jobslot was occupied when
assigning new job-id's for a particular worker. That would cause all
sorts of problems when one wants to check which jobs haven't finished
in a timely manner and could ultimately result in jobs going lost.

This patch corrects it by traversing the joblist in search for an
empty slot. Since the joblists are usually large enough for us to
run out of work before we run out of list, we should most of the time
end up in the best-case scenario of finding the first attempted slot
freely available, so we avoid fancy algorithms and just do a simple
traversal with O(n) worst-case complexity.

Signed-off-by: Andreas Ericsson <ae@op5.se>