workers: Add some basic job counting to our worker code
It would be nifty to allow workers a maximum of configurable jobs
before it's forced to die and respawn. This would allow us to use
leaky features that are still useful, such as embedded perl.

In order to make such configuration possible, it's vital that we
keep track of how many jobs each worker has been assigned. We
also need to know how many are still running, so we can make sure
not to assign jobs to a worker that is already overloaded with
jobs.

Signed-off-by: Andreas Ericsson <ae@op5.se>