lib/runcmd: Shut up some false positives from valgrind
When we encounter an execvp() error or the nul string, we would
previously leak some few bytes of memory. Since those bytes were
effectually leaked in a program that exited very quickly after
leaking the memory, it doesn't really matter in the slightest,
but it did make it a lot harder to find the real leaks due to
the incessant spamming of false positives, so we remove the fake
leak so valgrind can shut up and be happy again.

Signed-off-by: Andreas Ericsson <ae@op5.se>