lib/worker: Handle job timeouts properly
The new scheduling queue api makes this rather simple. Since the
scheduling API adds things in a LIFO manner we see a small
error that checks sometimes timeout less than one second after
they're scheduled if multiple items are scheduled to run on the
same second. Perhaps we need to fix that later. Right now it doesn't
seem terribly important though, and it would mean adding things
to the scheduling queue would need sorting, so the add operation
would be slower than it currently is.

Signed-off-by: Andreas Ericsson <ae@op5.se>