lib/worker: Be a bit more polite when checking for completions
If we catch a signal while in wait4(), it may well be that a child
exited while we were waiting for one, triggering an EINTR in the
wait4() call. In that case, we simply redo the wait, but for a max
time of 1 second to avoid hanging indefinitely in case many children
exit while we're wait()'ing for them.

Signed-off-by: Andreas Ericsson <ae@op5.se>