core: Make worker-receiver understand error worker responses
When a job fails to run properly for some reason (such as a timeout,
for example), the worker will set an "errcode" in the response vector
which corresponds to an errno that makes sense for the encountered
error. In the case of timeouts, we'll set it to ETIME, which in
human language means "Timer expired".

This patch teaches the receiver of worker responses to understand
such key/value pairs in the response.

Signed-off-by: Andreas Ericsson <ae@op5.se>