lib/iocache: Allow delimiters of len 1
Disregarding all of delim_len doesn't work when we want to find the
first byte of it while it sits at the end of the buffer, since we'll
search one byte short of it it in those cases. This was masked by a
bug in kvvec2buf() and how we used it and was discovered by our tests
to run the code with a delimiter of len 1.

Signed-off-by: Andreas Ericsson <ae@op5.se>