lib/iocache: Fix off-by-one error in iocache_use_delim()
Previously we'd set ptr to beyond the end of the delimiter, causing
weird off-by-one errors when using delimiters with proper lengths.

The bug was masked by the fact that we in our call to kvvec2buf()
from send_kvvec() always overallocated by one byte. Since kvvec2buf()
always adds a nul delimiter after the last value and before the real
message delimiter, we must take care to fix that too.

Signed-off-by: Andreas Ericsson <ae@op5.se>