lib/iobroker: Error out during compile if supporting multiple systems
The iobroker library is supposed to use a single backend system to
handle polling of its sockets (epoll, poll or select), but previously
one could specify multiple options and they would all be compiled in,
resulting in a horrible mess of crap returned from the iobroker api.

This patch fixes it by causing a compiler error if the user provides
more than one backend system for his or her polling needs.

Signed-off-by: Andreas Ericsson <ae@op5.se>