lib/iobroker: Make a timeout of -1 work well for select as well
poll() and epoll() uses a timeout value of -1 to mean "poll until we
get some data or a signal". This didn't work too well for the select()
based polling. Instead it caused it to return -1 immediately with errno
set to EINVAL, since the timeval->tv_sec value would be too large for a
time_t. On 32-bit systems, this patch effectually does nothing, since
the timeout there would equal some 80-ish years, but for 64-bit ones
it means we can run iobroker_poll() with a timeout of -1 and get the
"wait indefinitely" behaviour even when the iobroker api uses select()
in the backend.

Signed-off-by: Andreas Ericsson <ae@op5.se>