core: Skip sleeping and use the time to poll instead
Since we're turning Nagios into an eventbased state-machine rather
than a scheduling engine that actively checks for completion of
everything (by scouring spoolfiles and checking internal buffers for
new external commands), it doesn't make sense to spend time sleeping
when we could use that time to poll instead. If we get no events
during the time we poll, we effectively sleep anyway, and if we do,
we shouldn't have been sleeping in the first place. Either way, it's
a big fat win for lower latency and less resource waste.

Signed-off-by: Andreas Ericsson <ae@op5.se>