core: Use the iobroker to check for external commands
Since we'll be polling for events far more frequently than we used to,
we won't be needing the separate thread to magically enlarge the
buffer for external commands. Should the need arise anyway, we can
just modify the code to spawn a helper process for it that feeds us
results through a socket instead.

This means that Nagios is now single-threaded, which removes one huge
complexity for all addons and modules that wish to modify Nagios'
internals while it's running. It also means there's no need to specify
the "command_check_interval" anymore, as commands will always be
reaped as quickly as possible.

Signed-off-by: Andreas Ericsson <ae@op5.se>