base/workers: Initiate err_output to NULL
If we fail to get any output on stdout, we would previously copy the
errout variable to the plugin output stash, but we failed to properly
initialize that variable to NULL, so in cases where we got neither
stdout output or stderr output we might have used uninitialized memory
for plugin output, leading to "sometimes undefined behaviour".

This patch fixes it.

Signed-off-by: Andreas Ericsson <ae@op5.se>