core: Remove weird switch() statement when scanning check result queue
First of all, we use stat(2) instead of lstat(2), so the supposed check
for symlinks is bogus. If it was a symlink, we'd either get ENOENT or
the stat info of the actual file.

It's also wrong to check st.st_mode & S_IFMT against S_IFREG and then
checking it again with the macro S_ISREG(), since they'll both always
evaluate to true.

Instead, we remove the case statement and only use a simple S_ISREG()
and ignore everything but ordinary files.

Signed-off-by: Andreas Ericsson <ae@op5.se>