lib/iocache: Add functions iocache_{add,write,send,sendto}()
It's nifty to be able to cache output in memory before flushing it,
in order to avoid tons of small writes. This patch does just that,
although testing is currently lacking.

We're not really using these functions anywhere right now, but they've
been sitting in my working tree far too long not to be committed now.

Signed-off-by: Andreas Ericsson <ae@op5.se>